requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alibabacloud-dysmsapi20170525==3.1.0
alibabacloud-tea==0.4.0
alibabacloud_credentials==0.3.6
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.2
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.12
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
blinker==1.8.2
boto3==1.35.92
botocore==1.35.92
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cos-python-sdk-v5==1.9.33
crcmod==1.7
cryptography==44.0.0
daphne==4.1.2
edge-tts==6.1.18
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.0.3
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.27.0
Hypercorn==0.17.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
mysqlclient==2.2.6
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pillow==11.0.0
priority==2.0.0
propcache==0.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyMySQL==1.1.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
service-identity==24.2.0
six==1.17.0
SQLAlchemy==2.0.36
sympy==1.13.1
taskgroup==0.0.0a4
tencentcloud-sdk-python==3.0.1297
threadpoolctl==3.5.0
timm==1.0.12
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uWSGI==2.0.28
Werkzeug==3.1.1
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.17.1
zope.interface==7.2
